extract multi pointer handler
