+++ src/services/PointerHandler.ts
@@ -4,7 +4,7 @@
  * This won't show anything.
  */
 
-import { AnimationFrameId, emptyPos, IPos, IPosPair, unixMs } from '../misc';
+import { AnimationFrameId, emptyPos, IPos, unixMs } from '../misc';
 
 export interface IPointerHandlerProps {
   containing?: boolean;
@@ -15,9 +15,6 @@ export interface IPointerHandlerProps {
   size?: number;
   width?: number;
   onLongPoint?: () => void;
-  onPinchEnd?: () => void;
-  onPinchMove?: (pair: IPosPair, originalPair: IPosPair) => void;
-  onPinchStart?: (pair: IPosPair) => void;
   onPointCancel?: () => void;
   onPointEnd?: () => void;
   onPointMove?: (pos: IPos, startedPos: IPos) => void;
@@ -35,8 +32,6 @@ export default class PointerHandler {
   protected pointStartedPos: IPos = emptyPos;
   protected pointStartedAt: unixMs = 0;
   protected tmLongPressing: AnimationFrameId = 0;
-  protected pinching = false;
-  protected pinchOriginalPos: IPosPair = [emptyPos, emptyPos];
 
   protected get containing () {
     const { containing } = this.props;
@@ -57,19 +52,6 @@ export default class PointerHandler {
 
   protected get longPressing () {
     return this.tmLongPressing !== 0;
-  }
-
-  protected get pinchOriginalCenter (): IPos {
-    const [p1, p2] = this.pinchOriginalPos;
-    return {
-      x: (p1.x + p2.x) / 2,
-      y: (p1.y + p2.y) / 2,
-    };
-  }
-
-  protected get pinchOriginalDistance () {
-    const [p1, p2] = this.pinchOriginalPos;
-    return Math.sqrt((p1.x - p2.x) ** 2 + (p1.y - p2.y) ** 2);
   }
 
   constructor (props: IPointerHandlerProps) {
@@ -164,13 +146,6 @@ export default class PointerHandler {
 
       const pos = this.getPos(event, 0);
       this.startPressing(pos);
-    } else if (this.pressing && numTouches === 2) {
-      event.preventDefault();
-      const pair: IPosPair = [
-        this.getPos(event, 0),
-        this.getPos(event, 1),
-      ];
-      this.startPinching(pair);
     }
   }
 
@@ -179,9 +154,6 @@ export default class PointerHandler {
     if (this.pressing && numTouches === 1) {
       const pos = this.getPos(event, 0);
       this.movePressing(pos);
-    } else if (this.pinching && numTouches === 2) {
-      const posPair: IPosPair = [this.getPos(event, 0), this.getPos(event, 1)];
-      this.movePinching(posPair);
     }
   }
 
@@ -191,9 +163,6 @@ export default class PointerHandler {
     }
     if (this.pressing) {
       this.stopPressing();
-    }
-    if (this.pinching) {
-      this.stopPinching();
     }
   }
 
@@ -312,39 +281,6 @@ export default class PointerHandler {
     const p0 = this.pointStartedPos;
     const distance = Math.max(Math.abs(p0.x - pos.x), Math.abs(p0.y - pos.y));
     return distance > this.moveThreshold;
-  }
-
-  protected startPinching (posPair: IPosPair) {
-    this.stopLongPressing();
-    this.cancelPressing();
-
-    this.pinching = true;
-    this.pinchOriginalPos = posPair;
-
-    if (this.props.onPinchStart) {
-      this.props.onPinchStart(posPair);
-    }
-  }
-
-  protected movePinching (posPair: IPosPair) {
-    if (!this.pinching) {
-      return;
-    }
-
-    if (this.props.onPinchMove) {
-      this.props.onPinchMove(posPair, this.pinchOriginalPos);
-    }
-  }
-
-  protected stopPinching () {
-    if (!this.pinching) {
-      return;
-    }
-
-    this.pinching = false;
-    if (this.props.onPinchEnd) {
-      this.props.onPinchEnd();
-    }
   }
 
   protected getPos (event: MouseEvent): IPos;
